+++ scripts/generate_accessors.py
@@ -176,6 +176,29 @@ def {field}(self: 'fst.FST') -> None:
     self._put_slice(None, 0, 'end', {field!r})
 '''.strip())
 
+            elif field == 'targets':  # SPECIAL CASE!!!
+                print(f'''
+@property
+def {field}(self: 'fst.FST') -> FSTView:
+    """`FST` accessor for `AST` field `{field}`."""
+
+    ast = self.a
+    ast.{field}  # noqa: B018
+
+    if ast.__class__ in _ASTS_LEAF_ASSIGN_OR_TARGETS:
+        return FSTView_Assign_targets(self, {field!r})
+
+    return FSTView(self, {field!r})
+
+@{field}.setter
+def {field}(self: 'fst.FST', code: Code | None) -> None:
+    self._put_slice(code, 0, 'end', {field!r})
+
+@{field}.deleter
+def {field}(self: 'fst.FST') -> None:
+    self._put_slice(None, 0, 'end', {field!r})
+'''.strip())
+
             elif field == 'decorator_list':  # SPECIAL CASE!!!
                 print(f'''
 @property
@@ -293,19 +316,32 @@ from typing import Union
 
 from . import fst
 
-from .asttypes import ASTS_LEAF_VAR_SCOPE_DECL, AST, FunctionDef, AsyncFunctionDef, ClassDef, TypeAlias
+from .asttypes import (
+    ASTS_LEAF_VAR_SCOPE_DECL,
+    AST,
+    Assign,
+    FunctionDef,
+    AsyncFunctionDef,
+    ClassDef,
+    TypeAlias,
+    _Assign_targets,
+)
+
 from .astutil import constant
 from .common import PYGE12, PYGE13
 from .code import Code
 
 from .view import (
     FSTView,
+    FSTView_Assign_targets,
     FSTView_decorator_list,
     FSTView_comprehension_ifs,
     FSTView_Global_Nonlocal,
     FSTView_kwd_attrs,
     FSTView_dummy,
 )
+
+_ASTS_LEAF_ASSIGN_OR_TARGETS = frozenset([Assign, _Assign_targets])
 '''.strip())
 
     cardinality = {}  # {'field': 1 means single element | 2 means list (3 means can be either) | 4 if is optional (for single), ...}